Add files via upload
leafTranningMobileNet file contains the training part of the project. It contains data preprocessing and data augmentation which is used for generating training dataset. This file needs to be connected to google drive with all the images from plant village obtained from kaggle. This file is responsible to train and give output as models in hdf and tflite format.